feat(core): extract shared module
