+++ AGENTS.md
@@ -12,6 +12,7 @@
 - New command `list-diamonds` lists all tasks with `diamond` status.
 - New command `list-tasks` prints all tasks with their statuses.
 - New command `stats` provides summary analytics across tasks.
+- Core logic extracted to `src/core.ts`; CLI implementations should import from this module.
 - `crystal:sync` keeps `crystallization.json` synchronized between the repo and `.vscode` for IDE integration.
 - Diamond rule: tasks reaching all KPI thresholds get `diamond` status.
 - Use `crystal:init` to "crystallize" any repository with a starter `crystallization.json`.
